+++ warthog/api.py
@@ -15,12 +15,6 @@ Publicly importable API for the Warthog client and library.
 """
 
 from .core import (
-    NodeActiveConnectionsCommand,
-    NodeDisableCommand,
-    NodeEnableCommand,
-    NodeStatusCommand,
-    SessionEndCommand,
-    SessionStartCommand,
     STATUS_DISABLED,
     STATUS_DOWN,
     STATUS_ENABLED)
@@ -55,12 +49,6 @@ from .exceptions import (
 
 __all__ = [
     # warthog.core
-    'NodeActiveConnectionsCommand',
-    'NodeDisableCommand',
-    'NodeEnableCommand',
-    'NodeStatusCommand',
-    'SessionEndCommand',
-    'SessionStartCommand',
     'STATUS_DISABLED',
     'STATUS_DOWN',
     'STATUS_ENABLED',