+++ doc/source/changes.rst
@@ -1,9 +1,12 @@
 Change Log
 ==========
 
+0.8.1 - 2014-12-22
+------------------
+* Fixed small documentation issues and changed change log dates.
 
-0.8.0 - Future release
-----------------------
+0.8.0 - 2014-12-22
+------------------
 * **Breaking change** - Changed errors raised by :class:`warthog.config.WarthogConfigLoader`
   to be subclasses of :class:`warthog.exceptions.WarthogConfigError` instead of using errors
   from the standard library (:class:`ValueError`, :class:`IOError`, :class:`RuntimeError`).